The pointwise and simultaneous confidence intervals in `rootogram()` now rely on the exact `PoissonBinomial()` distribution (now available in `distributions3` devel) rather than its binomial approximation. The dependency on distributions3 devel was added to the DESCRIPTION.
In addition to pointwise and simultaneous confidence intervals for
`rootogram()`, `"tukey"` confidence intervals are now available which simply
correspond to limits of -1 and 1 for hanging or suspended rootograms. For
other flavors of rootograms these limits are transformed correspondingly.

Actually, "tukey" intervals are currently the only available option in the
ggplot2 interface because the PoissonBinomial information is not passed
on correctly at the moment.

The c() and print() method needed to be adapted for the additional
"distribution" column in rootogram objects. Tests have been extended
correspondingly.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/topmodels@1798 edb9625f-4e0d-4859-8d74-9fd3b1da38cb